requirements.txt
Django==4.2.9
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.1.0
django-imagekit==5.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-storages==1.14.2
boto3==1.34.69
